requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
imageio-ffmpeg==0.5.1
multidict==6.1.0
numpy==2.1.3
propcache==0.2.0
psutil==6.1.0
pycord==0.1.1
python-dotenv==1.0.1
vcgencmd==0.1.1
yarl==1.17.1
